[KVCache] pluggable kvcache spec
